Iskowela/requirements.txt
boto3>=1.9.96
botocore>=1.12.96
certifi>=2018.10.15
chardet>=3.0.4
dj-database-url>=0.5.0
Django>=3.0.2
django-crispy-forms>=1.8.1
crispy-bootstrap4>=2022.1
django-heroku>=0.3.1
django-storages>=1.7.1
docutils>=0.14
geoip2>=4.6.0
gunicorn>=20.0.4
idna>=2.6
jmespath>=0.9.3
Pillow>=7.0.0
python-dateutil>=2.6.1
pytz>=2019.3
requests>=2.18.4
s3transfer>=0.2.0
six>=1.11.0
whitenoise>=4.1.2